backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
twilio==7.0.0
requests==2.26.0
scikit-learn==0.24.2
numpy==1.21.2
pandas==1.3.3
python-jose==3.3.0
bcrypt==3.2.0
flask-socketio==5.1.1 